rename and move some files
