+++ client/src/app.tsx
@@ -1,10 +1,10 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import Welcome from './app/welcome/welcome';
 import DashboardLayout from './app/dashboard/layout';
-import ErrorPage from './Misc/ErrorPage';
-import AuthProvider from './Misc/AuthProvider';
-import AuthRoute from './Misc/AuthRoute';
-import NoAuthRoute from './Misc/NoAuthRoute';
+import ErrorPage from './components/error-page';
+import AuthProvider from './components/auth-provider';
+import AuthRoute from './components/auth-route';
+import NoAuthRoute from './components/no-auth-route';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 const router = createBrowserRouter([
