+++ client/src/components/auth-provider.tsx
