+++ client/src/components/auth-route.tsx
@@ -1,9 +1,10 @@
-import { useContext } from "react";
-import { AuthContext } from "./AuthProvider";
-import PropTypes from "prop-types";
-import { Navigate } from "react-router-dom";
+/* eslint-disable @typescript-eslint/strict-boolean-expressions */
+import { useContext } from 'react';
+import { AuthContext } from './auth-provider';
+import PropTypes from 'prop-types';
+import { Navigate } from 'react-router-dom';
 
-const AuthRoute = ({ children }: {children: any}) => {
+const AuthRoute = ({ children }: { children: any }): JSX.Element => {
   const { loading, user } = useContext<any>(AuthContext);
 
   if (loading) {
@@ -16,8 +17,6 @@ const AuthRoute = ({ children }: {children: any}) => {
 
   return <Navigate to="/" />;
 };
-
-
 
 AuthRoute.propTypes = {
   children: PropTypes.node,
