+++ client/src/components/no-auth-route.tsx
@@ -1,11 +1,12 @@
-import { useContext } from "react";
-import { AuthContext } from "./AuthProvider";
-import PropTypes from "prop-types";
-import { Navigate } from "react-router-dom";
+import { useContext } from 'react';
+import { AuthContext } from './auth-provider';
+import PropTypes from 'prop-types';
+import { Navigate } from 'react-router-dom';
 
-const NoAuthRoute = ({ children }: {children: any}) => {
+const NoAuthRoute = ({ children }: { children: any }): JSX.Element => {
   const { loading, user } = useContext<any>(AuthContext);
 
+  // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
   if (loading) {
     return <span className="loading loading-dots loading-lg"></span>;
   }
